created class ProjectConfig and moved class ErrorDetails
