+++ src/main/java/com/alexx/employeesAndDepartments/EmployeesAndDepartmentsApplication.java
@@ -1,20 +1,13 @@
 package com.alexx.employeesAndDepartments;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class EmployeesAndDepartmentsApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(EmployeesAndDepartmentsApplication.class, args);
-    }
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
     }
 
 }
