+++ src/main/java/com/alexx/employeesAndDepartments/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.alexx.employeesAndDepartments.openAPI;
+package com.alexx.employeesAndDepartments.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
