+++ src/main/java/com/alexx/employeesAndDepartments/config/ProjectConfig.java
@@ -1,0 +1,15 @@
+package com.alexx.employeesAndDepartments.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ProjectConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+
+}
